py_requirements.txt
google-cloud-bigquery-storage==2.27.0
google-cloud-bigquery==3.27.0
Flask==3.0.1
flask-restx==1.3.0
pandas==2.2.3
db-dtypes==1.3.1